Fixed content.
The following things need to be worked on,

4. Jointly hosted information has to be taken to the footer along with webmaster
5.  Other footer links to be included as in google site.
6.  About IIITDM Kurnool include the following

Indian Institute of Information Technology, Design and Manufacturing Kurnool (IIITDM Kurnool), fully funded by MHRD was announced in 2014 after receiving the assent of President of India to the Institutes of Information Technology Act, 2014. IIITDM Kurnool launched its academic program with the support of its mentor Institute, IIITDM Kancheepuram, in August 2015 from the Kancheepuram campus. Consequent to the approval of parliament, IIITDM Kurnool was accorded the status of the Institute of National Importance by making an amendment in the IIIT act 2014 on August 03, 2017.

7.  Add one para for IIIT Tiruchirapalli. (could combine into one)
8.  Make objectives as bullets without lines in between.  Do not divide objectives into two para.  Just one para suffices there.
9.  About the academy: make the current text into two para
10. Three menu items should be linked to three pages, registration, resources and upcoming events, respectively.
11. Registration page should contain two registrations
        1.  Click here to register for participating in a workshop
        2.  Click here to conduct a workshop.
12. In Resources page just include the link of Prof Srinivas in the brochure
In the upcoming events page make a table of three columns, sl no, event name, date and include
1, Our workshop name, MAY 2-3, 2020

13.  Give links to all three icons IIITK, IIITT, MHRD to corresponding websites
14.  Update link to IIIT Trichy menu item on the right side (not updated now)